requirements.txt
dj-database-url==2.2.0
Django==4.2.15
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djoser==2.2.3
environs==11.0.0
phonenumbers==8.13.43
pillow==10.4.0
psycopg2-binary==2.9.9